Fixed a warning on signed/unsigned comparison.
git-svn-id: http://blosc.pytables.org/svn/blosc/trunk@297 5b11b6af-1e7b-4319-8082-0aacbd7ffead

+++ blosc/blosc.c
@@ -532,7 +532,7 @@ int32_t compute_blocksize(int32_t clevel, uint32_t typesize, int32_t nbytes)
   uint32_t blocksize;
 
   /* Protection against very small buffers */
-  if (nbytes < typesize) {
+  if (nbytes < (int32_t)typesize) {
     return 1;
   }
